Fixed code warnings on smlua files (32-bit)

+++ src/pc/lua/smlua_cobject.c
@@ -84,7 +84,7 @@ static int smlua__get_field(lua_State* L) {
         return 0;
     }
 
-    u8* p = ((u8*)pointer) + data->valueOffset;
+    u8* p = ((u8*)(intptr_t)pointer) + data->valueOffset;
     switch (data->valueType) {
         case LVT_BOOL:       lua_pushboolean(L, *(u8* )p);              break;
         case LVT_U8:         lua_pushinteger(L, *(u8* )p);              break;
@@ -150,7 +150,7 @@ static int smlua__set_field(lua_State* L) {
         return 0;
     }
 
-    u8* p = ((u8*)pointer) + data->valueOffset;
+    u8* p = ((u8*)(intptr_t)pointer) + data->valueOffset;
     switch (data->valueType) {
         case LVT_BOOL:*(u8*) p = smlua_to_boolean(L, -1); break;
         case LVT_U8:  *(u8*) p = smlua_to_integer(L, -1); break;
